requirements.txt
asgiref==3.6.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==4.2
djangorestframework==3.14.0
idna==3.11
oauthlib==3.3.1
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
pytz==2023.3
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.4.3
#typing_extensions==4.5.0
urllib3==2.6.3
